[build-and-deploy] Stop running "rake clean"
This caused us to download initializr each time. It *used* to be
necessary because the downloaded initializr zip went into the "build"
directory.

However, in a commit or two ago I moved the zip to a tmp directory and
only move the extracted and needed files to the "build" directory.

By "build" directory I mean tmp/gh-pages.

By "a tmp directory" I mean tmp/initializr.